easy_parking/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytz==2024.1
setuptools==68.2.2
sqlparse==0.4.4
wheel==0.41.2
djangorestframework-simplejwt==5.3.1

alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-GraphQL==2.0.1
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
graphql-core==2.3.2
graphql-server-core==1.2.0
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.1.3
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.1
promise==2.3
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
Rx==1.6.3
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.17.0
Pillow>=10.2.0